pyproject.toml
[tool.poetry]
name = "computable-phenotypes"
version = "0.1.0"
description = ""
authors = ["anurag-bangera <97190957+anurag-bangera@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.2.1"
python-dotenv = "^1.0.1"
mysql-connector-python = "^8.3.0"
pyodbc = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
